Initial commit for the Schedule Meeting Modal Window
I modified the DateComponent to add the ability to highlight an array of dates.  I added a public route to the App component to
allow the modal window to be displayed.  I added an error function to the App component to handle server errors.  Seven files were added for the Modal Window function.